review_agents_gemini/pyproject.toml
[project]
name = "java-review-agent"
version = "0.1.0"
description = "Java Code Review AI Agent using LangGraph and Ollama"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.0.10",
    "langchain-community>=0.0.10",
    "ollama>=0.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
select = ["E", "F", "B", "I", "W"]
ignore = ["E501"]
line-length = 88

[tool.mypy]
strict = true
ignore_missing_imports = true
